# API/openapi.yaml
openapi: 3.0.3
info:
  title: ReleaseHub API
  version: 1.0.0
  description: Shared HTTP API contract for the ReleaseHub frontend and backend.
servers:
  - url: /
...
paths:
  /api/v1/access/capabilities:
    get:
      operationId: getAccessCapabilities
      summary: Get server-owned Access collection capabilities
      tags: [Access]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Access collection capabilities, content: { application/json: { schema: { $ref: "#/components/schemas/AccessCapabilitiesResponse" } } } }
        "404": { description: The caller has no manageable Access scope. }
  /api/v1/access/users:
    get:
      operationId: listAccessUsers
      summary: List manageable users
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string, maxLength: 128 } }
        - { $ref: "#/components/parameters/AccessStatus" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: One user page, content: { application/json: { schema: { $ref: "#/components/schemas/AccessUserListResponse" } } } }
        "404": { description: The caller has no manageable Access scope. }
    post:
      operationId: createAccessUser
      summary: Create a local user with a first-login password change requirement
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateAccessUserRequest" } } }
      responses:
        "201": { description: The local user was created, content: { application/json: { schema: { $ref: "#/components/schemas/AccessUserResponse" } } } }
        "404": { description: The caller lacks platform-management permission. }
        "409": { description: The username already exists. }
  /api/v1/access/snapshot:
    get:
      operationId: getAccessManagementSnapshot
      summary: Get access-control data manageable by the current user
      tags: [Access]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Platform administrators receive all data; Project managers receive only data within manageable scopes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessManagementSnapshotResponse"
        "401":
          description: The user is unauthenticated or the session has expired.

        "404":
          description: Access-management data is unavailable or the caller lacks management permission.
  /api/v1/access/groups:
    get:
      operationId: listAccessGroups
      summary: List manageable Groups
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string, maxLength: 128 } }
        - { $ref: "#/components/parameters/AccessStatus" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: One Group page, content: { application/json: { schema: { $ref: "#/components/schemas/AccessGroupListResponse" } } } }
        "404": { description: The caller has no manageable Access scope. }
    post:
      operationId: createAccessGroup
      summary: Create a Group
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateAccessGroupRequest" } } }
      responses:
        "201": { description: The Group was created, content: { application/json: { schema: { $ref: "#/components/schemas/AccessGroupResponse" } } } }
        "404": { description: The owner does not exist or the caller lacks management permission. }
        "409": { description: The name conflicts or the current data state does not permit creation. }
...
  /api/v1/access/groups/{groupId}/membership-candidates:
    get:
      operationId: listAccessMembershipCandidates
      summary: Find minimal user identities eligible for one manageable Group
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: groupId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: query, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 20, default: 20 } }
      responses:
        "200": { description: Eligible minimal user identities, content: { application/json: { schema: { $ref: "#/components/schemas/AccessUserCandidatesResponse" } } } }
        "404": { description: The Group does not exist or the caller lacks management permission. }
  /api/v1/access/scopes/{scopeKind}/{scopeId}/options:
    get:
      operationId: getAccessScopeOptions
      summary: Get legal Group, Role, and Permission options for one selected scope
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: scopeKind, in: path, required: true, schema: { type: string, enum: [platform, project, environment, application] } }
        - { name: scopeId, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        "200": { description: Legal options for the selected scope, content: { application/json: { schema: { $ref: "#/components/schemas/AccessScopeOptionsResponse" } } } }
        "404": { description: The scope does not exist or the caller lacks management permission. }
  /api/v1/access/roles:
    get:
      operationId: listAccessRoles
      summary: List manageable Roles
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string, maxLength: 128 } }
        - { $ref: "#/components/parameters/AccessStatus" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: One Role page, content: { application/json: { schema: { $ref: "#/components/schemas/AccessRoleListResponse" } } } }
        "404": { description: The caller has no manageable Access scope. }
    post:
      operationId: createAccessRole
      summary: Create a custom Role
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateAccessRoleRequest" } } }
      responses:
        "201": { description: The Role was created, content: { application/json: { schema: { $ref: "#/components/schemas/AccessRoleResponse" } } } }
        "404": { description: "The owner or permission does not exist, or the caller lacks management permission." }
        "409": { description: The name conflicts or the current data state does not permit creation. }
  /api/v1/access/bindings:
    get:
      operationId: listAccessBindings
      summary: List manageable Role bindings
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/AccessStatus" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: One Role binding page, content: { application/json: { schema: { $ref: "#/components/schemas/AccessBindingListResponse" } } } }
        "404": { description: The caller has no manageable Access scope. }
    post:
      operationId: createAccessBinding
      summary: Bind a Group and Role to a scope
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateAccessBindingRequest" } } }
      responses:
        "201": { description: The Role binding was created, content: { application/json: { schema: { $ref: "#/components/schemas/AccessBindingResponse" } } } }
        "404": { description: "The Group, Role, or scope does not exist, or the caller lacks management permission." }
        "409": { description: The binding conflicts or the current data state does not permit creation. }
  /api/v1/access/denies:
    get:
      operationId: listAccessDenies
      summary: List manageable explicit Deny policies
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/AccessStatus" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: One Deny policy page, content: { application/json: { schema: { $ref: "#/components/schemas/AccessDenyListResponse" } } } }
        "404": { description: The caller has no manageable Access scope. }
    post:
      operationId: createAccessDeny
      summary: Create an explicit Deny policy
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateAccessDenyRequest" } } }
      responses:
        "201": { description: The Deny policy was created, content: { application/json: { schema: { $ref: "#/components/schemas/AccessDenyResponse" } } } }
        "404": { description: "The Group, permission, or scope does not exist, or the caller lacks management permission." }
        "409": { description: The Deny policy conflicts or the current data state does not permit creation. }
...
  /api/v1/access/memberships:
    get:
      operationId: listAccessMemberships
      summary: List manageable Group memberships
      tags: [Access]
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: groupId, in: query, schema: { type: string, format: uuid } }
        - { name: userId, in: query, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/AccessStatus" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: One Membership page, content: { application/json: { schema: { $ref: "#/components/schemas/AccessMembershipListResponse" } } } }
        "404": { description: The caller has no manageable Access scope. }
...
components:
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    AccessStatus:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [all, active, inactive]
        default: all
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: releasehub_session
  schemas:
...
    AccessUser:
      type: object
      required: [id, username, disabled, allowedActions]
      properties:
        id: { type: string, format: uuid }
        username: { type: string, minLength: 1 }
        disabled: { type: boolean }
        allowedActions: { type: array, items: { $ref: "#/components/schemas/AccessAction" } }
    AccessGroup:
      type: object
      required: [id, ownerKind, name, oidcViewerOnly, disabled, allowedActions]
      properties:
        id: { type: string, format: uuid }
        ownerKind: { type: string, enum: [platform, organization, project] }
        ownerId: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1 }
        systemKey: { type: string, nullable: true }
        oidcViewerOnly: { type: boolean }
        disabled: { type: boolean }
        allowedActions: { type: array, items: { $ref: "#/components/schemas/AccessAction" } }
    AccessRole:
      type: object
      required: [id, ownerKind, name, active, permissions, allowedActions]
      properties:
        id: { type: string, format: uuid }
        ownerKind: { type: string, enum: [platform, project] }
        ownerId: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1 }
        systemKey: { type: string, nullable: true }
        active: { type: boolean }
        permissions:
          type: array
          items: { type: string }
        allowedActions: { type: array, items: { $ref: "#/components/schemas/AccessAction" } }
    AccessPermission:
      type: object
      required: [key, platformOnly, projectRoleDelegable]
      properties:
        key: { type: string }
        platformOnly: { type: boolean }
        projectRoleDelegable: { type: boolean }
    AccessMembership:
      type: object
      required: [id, groupId, groupName, userId, username, source, active, allowedActions]
      properties:
        id: { type: string, format: uuid }
        groupId: { type: string, format: uuid }
        groupName: { type: string }
        userId: { type: string, format: uuid }
        username: { type: string }
        source: { type: string, enum: [manual, oidc] }
        active: { type: boolean }
        allowedActions: { type: array, items: { $ref: "#/components/schemas/AccessAction" } }
    AccessBinding:
      type: object
      required: [id, groupId, groupName, roleId, roleName, scopeKind, active, allowedActions]
      properties:
        id: { type: string, format: uuid }
        groupId: { type: string, format: uuid }
        groupName: { type: string }
        roleId: { type: string, format: uuid }
        roleName: { type: string }
        organizationId: { type: string, format: uuid, nullable: true }
        scopeKind: { type: string, enum: [platform, project, environment, application] }
        projectId: { type: string, format: uuid, nullable: true }
        environmentId: { type: string, format: uuid, nullable: true }
        applicationId: { type: string, format: uuid, nullable: true }
        active: { type: boolean }
        allowedActions: { type: array, items: { $ref: "#/components/schemas/AccessAction" } }
    AccessDeny:
      type: object
      required: [id, groupId, groupName, permission, organizationId, scopeKind, projectId, active, allowedActions]
      properties:
        id: { type: string, format: uuid }
        groupId: { type: string, format: uuid }
        groupName: { type: string }
        permission: { type: string }
        organizationId: { type: string, format: uuid }
        scopeKind: { type: string, enum: [project, environment, application] }
        projectId: { type: string, format: uuid }
        environmentId: { type: string, format: uuid, nullable: true }
        applicationId: { type: string, format: uuid, nullable: true }
        active: { type: boolean }
        allowedActions: { type: array, items: { $ref: "#/components/schemas/AccessAction" } }
    AccessAction:
      type: string
      enum: [disable, addMember, viewMemberships, revoke]
    AccessCollectionCapability:
      type: object
      required: [key, visible, canCreate]
      properties:
        key: { type: string, enum: [users, groups, roles, memberships, bindings, denies] }
        visible: { type: boolean }
        canCreate: { type: boolean }
    AccessCapabilities:
      type: object
      required: [collections, permissions]
      properties:
        collections: { type: array, items: { $ref: "#/components/schemas/AccessCollectionCapability" } }
        permissions: { type: array, items: { $ref: "#/components/schemas/AccessPermission" } }
    AccessCapabilitiesResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessCapabilities" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    AccessUserListResponse:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AccessUser" } }
        meta: { $ref: "#/components/schemas/CursorPageMeta" }
    AccessGroupListResponse:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AccessGroup" } }
        meta: { $ref: "#/components/schemas/CursorPageMeta" }
    AccessRoleListResponse:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AccessRole" } }
        meta: { $ref: "#/components/schemas/CursorPageMeta" }
    AccessMembershipListResponse:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AccessMembership" } }
        meta: { $ref: "#/components/schemas/CursorPageMeta" }
    AccessBindingListResponse:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AccessBinding" } }
        meta: { $ref: "#/components/schemas/CursorPageMeta" }
    AccessDenyListResponse:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AccessDeny" } }
        meta: { $ref: "#/components/schemas/CursorPageMeta" }
    AccessUserCandidatesResponse:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AccessUser" } }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    AccessScopeOptions:
      type: object
      required: [groups, roles, permissions]
      properties:
        groups: { type: array, items: { $ref: "#/components/schemas/AccessGroup" } }
        roles: { type: array, items: { $ref: "#/components/schemas/AccessRole" } }
        permissions: { type: array, items: { $ref: "#/components/schemas/AccessPermission" } }
    AccessScopeOptionsResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessScopeOptions" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    AccessManagementSnapshot:
      type: object
      required: [canManagePlatform, managedProjectIds, users, groups, roles, permissions, memberships, bindings, denies]
      properties:
        canManagePlatform: { type: boolean }
        managedProjectIds: { type: array, items: { type: string, format: uuid } }
        users: { type: array, items: { $ref: "#/components/schemas/AccessUser" } }
        groups: { type: array, items: { $ref: "#/components/schemas/AccessGroup" } }
        roles: { type: array, items: { $ref: "#/components/schemas/AccessRole" } }
        permissions: { type: array, items: { $ref: "#/components/schemas/AccessPermission" } }
        memberships: { type: array, items: { $ref: "#/components/schemas/AccessMembership" } }
        bindings: { type: array, items: { $ref: "#/components/schemas/AccessBinding" } }
        denies: { type: array, items: { $ref: "#/components/schemas/AccessDeny" } }
    AccessManagementSnapshotResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessManagementSnapshot" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    CreateAccessGroupRequest:
      type: object
      required: [ownerKind, name, oidcViewerOnly]
      properties:
        ownerKind: { type: string, enum: [platform, organization, project] }
        ownerId: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1, maxLength: 128 }
        oidcViewerOnly: { type: boolean }
    CreateAccessUserRequest:
      type: object
      required: [username, initialPassword]
      properties:
        username: { type: string, minLength: 1, maxLength: 128, pattern: "^[^\\s]+$" }
        initialPassword: { type: string, minLength: 8, maxLength: 72, format: password }
    AccessUserResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessUser" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    CreateAccessRoleRequest:
      type: object
      required: [ownerKind, name, permissions]
      properties:
        ownerKind: { type: string, enum: [platform, project] }
        ownerId: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1, maxLength: 128 }
        permissions: { type: array, minItems: 1, uniqueItems: true, items: { type: string } }
...
    AccessScopeInput:
      type: object
      required: [scopeKind]
      properties:
        organizationId: { type: string, format: uuid, nullable: true }
        projectId: { type: string, format: uuid, nullable: true }
        scopeKind: { type: string, enum: [platform, project, environment, application] }
        environmentId: { type: string, format: uuid, nullable: true }
        applicationId: { type: string, format: uuid, nullable: true }
    CreateAccessBindingRequest:
      allOf:
        - $ref: "#/components/schemas/AccessScopeInput"
        - type: object
          required: [groupId, roleId]
          properties:
            groupId: { type: string, format: uuid }
            roleId: { type: string, format: uuid }
    CreateAccessDenyRequest:
      allOf:
        - $ref: "#/components/schemas/AccessScopeInput"
        - type: object
          required: [groupId, permission]
          properties:
            groupId: { type: string, format: uuid }
            permission: { type: string }
    AccessGroupResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessGroup" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    AccessRoleResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessRole" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
...
    AccessBindingResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessBinding" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    AccessDenyResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/AccessDeny" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    ResponseMeta:
      type: object
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
          minLength: 1
        timestamp:
          type: string
          format: date-time
...
    CursorPageMeta:
      type: object
      required:
        - requestId
        - timestamp
        - hasMore
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        nextCursor:
          type: string
        hasMore:
          type: boolean
...